assignment_05/pyproject.toml
[project]
name = "assignment-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
]
